Fix handling `NaN` values when fitting JS univariate drift (#340)
* Add column & method for univariate fitting errors

* Refactor to use single data cleaning method

* Filter NaN's when fitting JS

* Refactor data cleaning to accept columns argument

Previously the data cleaning method operated by accepting multiple
dataframes and inspecting each dataframe separetely for `NaN`'s.
Depending on how the data is processed after cleaning, splitting columns
into separate dataframes can be rather annoying.

To avoid that this commit changes the method to accept a single
dataframe and a columns argument. The columns argument specifies which
column subsets should be inspected for `NaN`'s, enabling the same
behaviour using a more convenient syntax.

* Remove errors and use warning behaviour instead

The performance calculator for binary classification had checks in place
to generate an exception if the prediction column contains nothing but
`NaN`'s. This behaviour contradicts the warning functionality that is in
the same functions that would should return `NaN` and issue a warning.
It is also inconsistent with other calculators which do issue a warning
instead of raising an error.

This commit removes the errors and relies on the existing warning
functionality.

* Refactor more data cleaning methods

* Deal with mypy overload issue

---------

Co-authored-by: Niels Nuyttens <niels@nannyml.com>