Remove chi2 thresholds for analysis & reference (#349)
Chi2 thresholding is based on p-values, which means the calculated
thresholds are irrelevant. Previously thresholds were removed as part of
alerting. As this is only applied on analysis data the threshold value
was still available on reference data, which led to confusion.

This commit changes that so the chi2 threshold values are removed across
the entire dataset.