Fix plotting confidence bands with gaps
Plotly doesn't correctly handle fills with gaps. This is an issue for
confidence bands as they are being plotted currently. Work around is to
tell Plotly to fill the gaps, which causes correct rendering.
Unfortunately it does fill the gap, but this is the easiest work around
for now.

See https://github.com/plotly/plotly.js/issues/2736 for context.