Update analytics for compatibility with cloud SDK (#376)
The `analytics-python` package currently used has a dependency on
backoff `==1.10.0`. This conflicts with the NannyML Cloud SDK which has
a dependency on backoff `>=1.11.1,<3.0` (because of its `gql`
dependency).

This commit fixes that by updating the analytics package to the latest
version. It was renamed to `segment-analytics-python` but still performs
the same package as the original `analytics-python`.