Optimize filtering univariate result for period
The `PerMetricPerColumnResult` filter function has a high overhead for
selecting a subset of columns or metrics. This overhead is also incurred
(and highest) when only filtering for period, as then all columns &
metrics will be selected.

This commit adds a short-circuit path to avoid the overhead when only
the period requires filtering. For a result with 50 columns and 8
metrics this results in a >100x speed-up when only filtering for period.